getting delete function working reverted wagonSetList back to an ArrayList cleaned up naming error added toString to most data classes for nicer reading during deletion
note:
a lot of debugging is required. the deletion buttons show error when an item is deleted and its part of some other object (for example deleting a platform that is being used by a journey)
oddly enough it brings up the error but still does as expected
I will work on this tomorrow as it is getting late.